refactor(edge): isolate bootstrap outcome recording and metrics wiring
